requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
av==10.0.0
babel==2.16.0
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.12.2
bitarray==2.9.2
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
cinemagoer==2023.5.1
classifier==2.0
click==8.1.3
cloudpathlib==0.16.0
cloudpickle==3.0.0
cmake==3.27.5
colorama==0.4.6
colorlog==6.8.2
comm==0.1.3
comtypes==1.2.0
confection==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
cymem==2.0.8
cytoolz==0.12.3
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
DateTime==5.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
demucs==4.0.1
distlib==0.3.7
dlib==19.24.2
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
dora_search==0.1.12
EasyProcess==1.1
ecapture==2.0.2
einops==0.7.0
emoji==2.11.0
encodec==0.1.1
entrypoint2==1.1
eth-account==0.11.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
executing==1.2.0
fastapi==0.109.2
fastjsonschema==2.20.0
ffmpy==0.3.2
filelock==3.12.4
filetype==1.2.0
flashy==0.0.2
Flask==2.2.2
flatbuffers==24.3.25
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
gradio==3.50.2
gradio_client==0.6.1
graphviz==0.20.3
greenlet==2.0.2
grpcio==1.59.0
gTTS==2.3.2
h11==0.14.0
h5py==3.11.0
hexbytes==0.3.1
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
ifaddr==0.2.0
IMDbPY==2022.7.9
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.22.0
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
julius==0.2.7
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
keyring==24.2.0
kiwisolver==1.4.5
lameenc==1.7.0
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-text-splitters==0.0.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.40
lazy_loader==0.3
lib==4.0.0
libclang==16.0.6
libmagic==1.0
librosa==0.10.1
lightning-utilities==0.10.1
llvmlite==0.42.0
lru-dict==1.2.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.21.1
matplot==0.1.9
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.5
mistune==3.0.2
ml-dtypes==0.4.0
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
mss==9.0.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.1
nh3==0.2.14
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
num2words==0.5.13
numba==0.59.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.0.76
opencv-python-headless==4.10.0.84
openunmix==1.2.1
opt-einsum==3.3.0
optree==0.12.1
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.1
parsimonious==0.9.0
parso==0.8.3
pathlib_abc==0.1.1
pathy==0.11.0
pickleshare==0.7.5
Pillow==10.0.1
pipreq==0.4
pipreqs==0.5.0
pkginfo==1.9.6
platformdirs==3.10.0
plotly==5.24.1
pooch==1.8.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.14
pydantic_core==2.16.2
pydeck==0.8.1b0
pydot==3.0.3
pydub==0.25.1
pyee==11.0.0
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.14.0
pyjokes==0.6.0
PyJWT==2.8.0
pylibsrtp==0.8.0
pyloco==0.0.139
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
pyproject-toml==0.0.10
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-iso639==2024.2.7
python-json-logger==2.0.7
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3.post1
pyunormalize==15.1.0
pywhatkit==5.4
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.0.2
rapidfuzz==3.7.0
readme-renderer==42.0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.5.3
rlp==4.0.0
rpds-py==0.10.3
rsa==4.9
ruff==0.2.1
safetensors==0.4.2
scikit-learn==1.6.0
scipy==1.11.3
scraper-api==0.1.0
secure-smtplib==0.1.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
shellingham==1.5.4
SimpleWebSocketServer==0.1.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
spotify==0.10.2
SQLAlchemy==2.0.21
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.36.3
streamlit==1.31.1
streamlit-chat==0.1.1
streamlit-webrtc==0.47.1
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.1
tensorflow==2.17.0
tensorflow-estimator==2.14.0
tensorflow-hub==0.16.1
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-serving-api==2.5.0
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
tf_keras==2.17.0
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.3.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.3.1
torchvision==0.15.2
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.37.2
treetable==0.2.5
twilio==8.9.0
twine==4.0.2
typer==0.7.0
types-python-dateutil==2.9.0.20240821
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.13.2
unstructured-client==0.18.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.5
ushlex==0.99.1
uvicorn==0.27.1
validators==0.22.0
virtualenv==20.24.5
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
weasel==0.3.4
web3==6.15.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.2.2
widgetsnbextension==4.0.13
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.14.1
xformers==0.0.22
xmltodict==0.13.0
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
zope.interface==6.0
